Fix int type

+++ Word-Flow/adjmenu.cpp
@@ -58,7 +58,7 @@ void AdjMenu::on_dic_show_Button_clicked()
 
     while (query.next())
     {
-        if ((query.value(0) > 100) && (query.value(0) < 201)){
+        if ((query.value(0).toInt() > 100) && (query.value(0).toInt() < 201)){
             QString english = query.value(1).toString();
             QString russian = query.value(2).toString();
             ui->textBrowser->insertPlainText(english+" - "+russian+"\n");
